backend/rest-api/register-user/requirements.txt
bcrypt==4.0.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
psycopg2-binary==2.9.3
python-dotenv==1.0.0
Werkzeug==2.2.2
